// heart-watch-heartbeat/src/types/prediction.ts
export interface PredictionInput {
  age: number;
  sex: number;
  cp: number;
  trestbps: number;
  chol: number;
  fbs: number;
  restecg: number;
  thalach: number;
  exang: number;
  oldpeak: number;
  slope: number;
  ca: number;
  thal: number;
}

export interface PredictionResult {
  probability: number;
  risk: "Low" | "Moderate" | "High";
}